All the changes which successfully parse and replicate the reference scenario at least.

+++ cvs/objects/ccarbon_model/include/land_carbon_densities.h
@@ -106,7 +106,7 @@ protected:
         DEFINE_VARIABLE( SIMPLE, "below-ground-carbon-density", mAvgBelowGroundCarbon, double ),
         
         //! Age at maturity.  This is used to grow forests slowly.
-        DEFINE_VARIABLE( SIMPLE, "mature-age", mMatureAge, int )
+        DEFINE_VARIABLE( SIMPLE, "mature-age", mMatureAge, /*int*/double )
     )
 };
